Added options "-c" to use the Codex model and the option "-m" to allow multiline text.
